Always use FQDN for android-git
Using the FQDN matches what we tell people to use when they invoke
'repo init -u ...', so individual projects will use the same URL
as the manifest uses.  This means that repo can reuse the same
SSH connection for both the manifest and the projects, slightly
improving our sync times.