Switch Market branch to market-myapps
For this weekend HC build, we'd like to go ahead and switch the
version of Market bundled with the system image to market-myapps.
We created the market-myapps branch this week to start work on
the headless version of the market that will not include any of
the browsing, purchasing, billing functionality that's now to be
handled by Finsky.

Testing for this version of Market shows that Finsky- and George-
initiated app purchases and downloads work, and licensing
functionality is preserved.

Bug: 3194829