Let uiautomator have its own branches
uiautomator has been using revision="master" in the manifest.  Recently
there was a change to doclava and checkapi (see bug 11293324), which
are not pinned to revision="master", which broke the build because
uiautomator interacts with doclava.  Ed has pushed uiautomator at
'1640a21f' (which is the version right before the change that broke
klp-dev) to all of the klp* branches.  This change, by removing
revision="master" from the manifest will, let klp-dev (and all of the
klp branches that auto-merge the manifest from klp-dev) switch from
revision="master" (which has the breaking CL, that is needed in master)
to instead have their own branch of uiautomator (which should work
in the klp* branches as it doesn't have the change that went into master)

Bug: 13475771
Change-Id: I9c83427501a93dc6bb403f61799d278d0297af0d
Reviewed-on: https://googleplex-android-review.git.corp.google.com/434017
Reviewed-by: Hui Shu <hush@google.com>
Reviewed-by: Ed Heyl <edheyl@google.com>