Change owners lib to use the antman branch instead of master
We are using GmsCore Jarlsberg prebuilt in master. so we need to use
antman branch.
Once we switch to Kenafa prebuilt in master, we will be able to switch to
batman branch and/or revert to master branch of owners

Second CL as I wrote batman instead of antman in the first one

Change-Id: I5fdfab0c454e12967cf90f754f195f2403beb353
Reviewed-on: https://googleplex-android-review.git.corp.google.com/435089
Reviewed-by: Nicolas Klein <foxykeep@google.com>