Remove external/v8
external/v8 (javascript compiler) was used by the classic WebView.  The new Chromium
WebView uses external/chromium_org/v8.  external/v8 is currently out of date and doesn't
work with arm64.  When we went to delete it (after moving to the Chromium WebView) we
found that chromium-libpac was also using it.  We have worked with chromium-libpac folks
and have gotten them off of external/v8.  I removed external/v8 from my ToT master tree
and did a "make clean" followed by a "make" and was able to build hammerhead-userdebug
with no errors.

Bug: 12452053
Change-Id: I02376e72fa04ddfafe731d81fd64a0a1dca3dac1
Reviewed-on: https://googleplex-android-review.git.corp.google.com/441217
Reviewed-by: Narayan Kamath <narayan@google.com>
Reviewed-by: Richard Coles <torne@google.com>
Reviewed-by: Daniel Cashman <dcashman@google.com>