Remove play-common from master.
It's 5 versions out of date, and no-one in master is using it.
The CL below removes the last dependency, which was only there
to transitively depend on golly.

To be submitted after this has merged:
https://googleplex-android-review.git.corp.google.com/#/c/609521/
Change-Id: I7235d7b5bf4209b40171c327c31cea193248661e
Reviewed-on: https://googleplex-android-review.git.corp.google.com/609482
Reviewed-by: Narayan Kamath <narayan@google.com>
Reviewed-by: Geoff Mendal <mendal@google.com>
(cherry picked from commit d7500a5a5c46116f7ea243175c31b6d01d957fdb)
Reviewed-on: https://googleplex-android-review.git.corp.google.com/613576
Reviewed-by: Jeff Davidson <jpd@google.com>
Reviewed-by: Ying Wang <wangying@google.com>