Put clang back in the manifest.
We still need this because we pull the analyzer scripts from here
(they aren't installed in the prebuilts directory). We'll want to move
to having it installed in prebuilts eventually, but the minimal fix is
fine for now.

This went unnoticed because failures from build-analyzer.sh do not
cause build-host-prebuilts.sh to fail (which I'll fix once this gets
merged).