Move build/ to build/make/
Remove the nesting of build/{blueprint,kati,soong}, which should allow
us to use submodules. Differs from master in that most of the
copyfile/linkfile entries have been removed, since we really only need
this repository for the tools in build/tools. Everything builds with
Soong or Go's scripts, not Make.

Bug: 28001743
Test: prebuilts/build-tools/build-prebuilts.sh
Change-Id: I5ff919cdb0423ebd11d9033fadab8aedf1a8b81d