Remove vendor/google from studio-master-dev manifest.
The platform/vendor/google project in master is being split (see
go/vendor_google_split). As part of that split we need to move anything we have
in that project to one of the specific projects in that go link. I looked in
that project and we don't have anything. There is a build_trigger directory
and there is a .git file in that directory. This CL will just remove
vendor/google from studio-master-dev checkout. Note, that this is different
than tools/vendor/google which contains adrt.

Bug: 68863888
Test: N/A
Change-Id: Ida010b79726708d865df15104b630aa94e337878