Remove gobo from master
The gobo device has been End Of Lifed from master and future
releases.  Its last "release" is PI, where it remains (although
it wasn't actually released on PI).

We remove:
  device/huawei/gobo
  device/huawei/gobo-kernel
  vendor/mediatek/mt6735m/*

Note that the gobo was the only device on master using the
MT6735 chipset, hence the removal of those projects as well.

Test: TreeHugger
Bug: 80133532
Change-Id: Iaa5b558c0c84f8ddf79d0a7a8b479892387157fe