Remove bazel.rc symlink after specifying in bazel wrapper (ag/5429817).
This reverts commit 8ef73596c4632387d46590cd8a6b1d2b89366127.

Reason for revert: This revert removes the symlink in WORKSPACE to
tools/base/bazel/toplevel.bazel.rc. This same change was reverted
when it was discovered that ab/studio-master-dev, studio target
does some bazel invocations initiated from ant and/or gradle. That
broke bazel because it could no longer find the bazel.rc which was
specified explicitly in the bazel recipe. Since then, the bazelrc
specification has been removed from the recipe and put into the
bazel wrapper. Currently, there are two bazel.rc files being passed
to every bazel invocation. The symlink and the one specified in the
wrapper. After this lands, bazel invocations will have one explicitly
passed bazel.rc which is specified in the wrapper.

Change-Id: I38989025b6e1a7e44a2be12af4ddf375f2f96e60