Pin external/toybox to android-q-preview-2.5 (2019-04-08) in master-art.
Recent revisions of Toybox in the AOSP tree seem to be the cause of
test failures on the ART
Buildbot (https://ci.chromium.org/p/art/g/luci/console), presumably in
`env`. Pin the revision of external/master to an earlier revision in
the master-art manifest to have the bot turn green back again while we
investigate.

Test: Rely on the ART Buildbot.
Bug: 131667225