Add clang
Since we aren't branching the clang prebuilts we pin to a specific
version in the manifest. This can be updated to any commit from
mirror-aosp-master if need be.
Note that currently there will be an issue using a more recent version
of the windows clang, since in later versions some of the symlinks
we're creating via repo are checked in, and currently when we sync on
rbe symlinks aren't created correctly (though they are created
correctly when added by repo).

Test: none
Bug: 135278528
Change-Id: I8de9a6e801a5797dd9ace895aac05240ef6a8e08