Pin the GCC manifest to the current mingw revision.
https://android-review.googlesource.com/c/platform/prebuilts/gcc/linux-x86/host/x86_64-w64-mingw32-4.8/+/1138078
updates mingw to use ucrt instead of msvcrt by default, which is
needed for binutils now that the gold plugin is built with ucrt (see
http://b/142337488).

Unfortunately GCC doesn't currently build with this change (sys_nerr
and sys_errlist are forward declared by libiberty, and they're defined
incorrectly for ucrt). We should look for upstream fixes or fix
locally, but for now, unblock the NDK release.

Test: None
Bug: http://b/142337488