Remove lottie from the default manifest.
It's not currently used, and upstream it putting out regular updates
that we're wasting time tracking.

It'll require a manifest change to re-enable anyway (because it now uses
okio, which hasn't gone through go/android-3p approval), so there's no
point keeping it in the manifest until/unless that happens.

Change-Id: Id116bec3809cc376bd84f29e8f502378651349ca