master-ndk: pin revision of a few gcc prebuilts
Rather than get these deprecated/to be removed binaries (GNU binutils)
to build with Clang, we're going to pin the revision of the GCC
prebuilts used to build them.  This will prevent emergency updates to
GCC/binutils from being used to rebuild these prebuilts, but the goal is
to use LLVM substitutes before then.

The SHAs correspond to the fixes to recently identified post submit
regressions for the `binutils` branch of the platform.

Pinning the revision will prevent the removal of GCC from breaking these
targets, until we stop building/using them outright.

Bug: 147240075

Change-Id: I060e38c7ec53253efa459a4f1a0c24ba47ac1f57
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>