Update protobuf prebuilts and source branch.
We add the protobuf 'prebuilts' repo which contains the officially
released protoc compiler, and switch both branches to the
'androidx-master-dev' branch which will contain the officially released
source code matching that compiler.

This is required to build the C++ proto lib in Jetpack with the source
code that exactly matches the protoc binary.

Bug: 157512302
Change-Id: I40b780296903ae660e6b6dbd94302e030c5b9c71