Pin googletest to a specific revision
A recent update to googletest is breaking our tests, because we were
referencing the master revision, which is always kept up to date.
For now, we resolve this by pinning to a version grabbed from a few
weeks ago, when things were still working.

The error was:

```
tools/base/profiler/native/perfd/cpu/trace_manager_test.cc:
  error: use of undeclared identifier 'DoAll'
  .WillOnce(DoAll(SaveArg<0>(&cmd_1), Return(true)))
            ^
```

Change-Id: Ic2e1e803a51b8a3818ff56312b31a3bc4d7b2ca5