Add lint tool used in Java compilation rules.
The Java rules add dependencies on a linter in prebuilts/cmdline-tools,
so with SOONG_ALLOW_MISSING_DEPENDENCIES we get missing dependency
errors. In a normal build those errors tend to end up on the lint
output and not the main jar outputs of the Java modules, but when
build/soong/scripts/build-aml-prebuilts.sh is used they do break the
whole module, and the missing dependency tracking in Soong doesn't
guarantee that is avoided.

Test: art/build/build-art-module.sh on master-art
Bug: 172480615
Change-Id: I6f0ba2cbcf727a40a8ffe1c243360241c96a6ab8