Enable repositories to enable root canal
This brings a series of repositories that enable us to start building
root canal as part of the emulator:

- system/bt this contains the whole bluetooth stack.

As part of the build root canal will compile pdl files to to header
files. In order for us to build this compiler we will need:

- flex, m4 and bison.

For the runtime we require:

- libchrome
- abseil

We update perfetto to a later version, to satisfy libchrome
dependencies.

Change-Id: I49ed6b9225a295ea187cf993233f04e4290ae7cc