Pin vendor/amlogic/common back to goog/master.
goog/master has been created with partner-android/master-tv-dev
at c0778db21809d5a5689d9a3990e30e9509010ecf.  This change is necessary
to make TH pass on the system/bt migration to
packages/modules/Bluetooth.  TH cannot take a change in a project which
is also referenced as changing in the manifest, making this step
necessary.

Original change was in topic:bug_196026708, but needs to be split out.

BUG: 196026708
Test: TH
Change-Id: Ia918b1d05a4397e34777c7b9cc562fe752b9b90d