Add non-apex dependencies of libbpf_android
Upcoming license changes use the existence of a dependency to track
license requirements.  libmeminfo depends on libbpf_android, and then
removes the dependency with target.apex.exclude_static_libs.  The apex
mutator runs after the deps mutator, which means the dependency already
exists before the apex variant is created, and is annotated to be
ignored in the apex variant.  The licensing code is unaware that the
dependency is ignored, causing it to traverse into libbpf_android
and hit missing dependency errors.  Add the dependencies of libbpf_android
as a workaround.

Bug: 207445310
Test: art/tools/buildbot-build.sh --target
Change-Id: I336aeb2062493021ffc2d73d195c629c0e35cf50