Make external/tink use emu-master-dev.
In earlier revision (aosp/1950539) the manifest was referencing
upstream-master, which is replicated from GitHub (google/tink).
This can cause issues in the future as the upstream was a replica from
google3 via copybara, but there is no presubmit validation in place to
prevent breaking our branch.

For now, create a emu-master-dev based on the latest upstream-master and
point the manifest back to the AOSP branch instead.

Bug: 230761320
Change-Id: I10a60857b10bcab485727942b35472bfd2f3a17e