Add projects for mokey device
See go/mokey for more details on the device.

We use 'clone-depth="2"' for several projects which are mostly
binaries (and possibly include files as well).  This prevents
folks from needing larger local histories of binaries on their
local disks, while still making it easy to locally revert a
single CL in case of a questionable merge.  We also add this
to a couple similar wembley projects to save space.

Note that we also switch the source of the vendor/mediatek/Android.mk
file from wembley sources into mokey sources.  The logic for this
file remains nearly identical.  But this move will allow the
removal of wembley source in the future to be a pure removal.

Bug: 258284421
Test: Verify that TreeHugger builds of all other targets for this CL don't have unexpected changes in artifact; Locally build with `lunch mokey-userdebug`
Change-Id: I417327dfce15effdf722406348637498ea1dd9a7