Remove the `vendor/google/tv/chromecast` project
`vendor/google/tv/chromecast` project is not used for anything in the
manifest of main Android dev branches. Further more its master (and
xyz-dev) revisions are completely unmaintained, as canonically this
projects lives in Partner Gerrit. This project is used from -tv-
branches where it tracks Partner Gerrit.

The current state is a remnant of an attempted, then aborted move of
this project from Partner to Internal Gerrit.

Change-Id: I749a16e884bd888209bd8d85308c91f3042c0a97
Bug: 266930988