Map "platform/build" to "build/make"
It currently is mapped to "build" and causes superproject / code search
issues as "build/bazel" is nested under it.

Android Build uses the "core/build_id.mk" in this project but references
it by project name "platform/build", not the mapped path so it should be
okay (go/ab-version#android-build-system).

Change-Id: Ia78f38be160ca567a0a0c3c198d5415754052a9c