Add dependency of aconfig to master-art
The dependencies were dumped from aosp/main with the following bazel-query
  bazel query --config=queryview 'deps(//build/make/tools/aconfig:aconfig--linux_glibc_x86_64)'
  bazel query --config=queryview 'deps(set(//external/rust/crates/protobuf-parse:libprotobuf_parse--linux_glibc_x86_dylib //external/rust/crates/regex:libregex--linux_glibc_x86_dylib))'

aconfig binary is required to build a Soong java_aconfig_library module.

I did successfully use the prebuilt aconfig in master-art, but the
verison is not latest, and we need the latest aconfig to unblock
go/mainline-on-main. Also, the build logic of using aconfig optionally is a
bit complicated as it involves other branches, and I haven't
figured out the best way to do so.

Bug: 316554963
Test: art/tools/buildbot-build.sh --dist succesfully with https://r.android.com/c/2880086/1
Change-Id: I0c4fa4d8baaece2cc5b9f10f6af926b619b0dac4