Add trusty/[device|user]/desktop to the android manifest
These projects need to be added to the manifest for Android desktop
work. Trusty is an AOSP-first project, and so work in Trusty for this
project is done AOSP-first as well.

The code location should be at //trusty/[device|user]/desktop

The device repo configures building the desktop trusty image.
The user repo holds Trusty userspace and Android library code that is
specific to Android desktop

Bug: 361760432
Flag: EXEMPT: manifest change
Test: 'repo sync --nmu --no-use-superproject' synced correctly
Change-Id: I95773fddc7786f9ea42e8eea82a53ef2f49a058f